+++ src/app/Components/ComponentsSamples/Alerting/Alerts/AlertSamples.xaml.cs
@@ -1,4 +1,5 @@
 using System.Windows.Input;
+using DIPS.Mobile.UI.Components.Alerting.Alert;
 using DIPS.Mobile.UI.Components.Alerting.SystemMessage;
 
 namespace Components.ComponentsSamples.Alerting.Alerts;
@@ -10,4 +11,12 @@ public partial class AlertSamples
         InitializeComponent();
     }
 
+    private void Button_OnClicked(object? sender, EventArgs e)
+    {
+        foreach (var vte in this.GetVisualTreeDescendants())
+        {
+            if(vte is AlertView alertView)
+                alertView.Animate();
+        }
+    }
 }